requirements.txt
<<<<<<< HEAD
beautifulsoup4==4.6.3
django-apscheduler==0.2.13
fbchat==1.4.2
python-telegram-bot==11.1.0
=======
beautifulsoup4 == 4.6.3
django-apscheduler == 0.2.13
fbchat == 1.4.2
python-telegram-bot == 11.1.0
>>>>>>> bfc488ad51715cf21e4e02fbe4765263f013a6aa
